py-backend/requirements.txt
fastapi
uvicorn
pydantic
boto3
botocore
typing-extensions
python-dotenv
langgraph
mcp
yfinance>=0.2.58
duckduckgo-search
setuptools
aiohttp
strands-agents
strands-agents-tools
python-docx
matplotlib
numpy